services/api/requirements.txt
# Web framework
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Auth
python-jose[cryptography]==3.3.0
bcrypt==4.2.1

# Config
pydantic==2.10.3
pydantic-settings==2.6.1

# LLM providers
anthropic==0.40.0
openai==1.57.0

# HTTP client
httpx==0.28.1

# Redis
redis[asyncio]==5.2.1

# Observability
sentry-sdk[fastapi]==2.19.2

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0

# Code quality
ruff==0.8.4
